# comments-service-openapi.yaml
openapi: 3.0.3
info:
  title: Comments Service API
  description: |
    The Comments Service manages comment lifecycle operations for blog posts including creation, 
    retrieval, moderation, and threading. This service operates independently and provides 
    RESTful endpoints for discussion management with support for nested commenting, moderation 
    workflows, and real-time updates.
    
    ## Features
    - CRUD operations for comments
    - Nested comment threading (up to 10 levels deep)
    - Comment voting system
    - Moderation and content filtering
    - Real-time updates via WebSocket
    - User reputation system
    - Comment search and filtering
    
    ## Threading System
    Comments support nested replies with a maximum depth of 10 levels. Each comment can have
    multiple replies, creating a tree structure that can be rendered in threaded or flat modes.
    
    ## Moderation
    Comments go through automatic and manual moderation processes including profanity filtering,
    spam detection, and user reputation checks.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
    url: https://docs.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4001
    description: Development server
  - url: https://api.example.com/comments
    description: Production server
security:
  - BearerAuth: []
  - {}
paths:
...
  /moderation/comments:
    get:
      summary: Get comments pending moderation
      description: |
        Returns comments that require moderation review. Only accessible to users 
        with moderator or admin roles.
      operationId: getModerationQueue
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by moderation status
          schema:
            type: string
            enum: [pending, flagged, reported]
            default: pending
        - name: priority
          in: query
          description: Filter by priority level
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Moderation queue retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Comment'
                        - type: object
                          properties:
                            moderationFlags:
                              type: array
                              items:
                                type: string
                            reportCount:
                              type: integer
                            priority:
                              type: string
                              enum: [low, medium, high, critical]
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /moderation/comments/{id}/action:
    post:
      summary: Take moderation action on a comment
      description: |
        Approve, reject, or take other moderation actions on a comment. 
        Only accessible to users with moderator or admin roles.
      operationId: moderateComment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Comment identifier
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationAction'
      responses:
        '200':
          description: Moderation action completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  commentId:
                    type: string
                    format: uuid
                  action:
                    type: string
                    enum: [approve, reject, flag, delete]
                  status:
                    type: string
                    enum: [active, moderated, deleted]
                  moderatedBy:
                    type: string
                    description: Moderator username
                  moderatedAt:
                    type: string
                    format: date-time
                  reason:
                    type: string
                    description: Moderation reason
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from the Authentication Service
  schemas:
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique comment identifier
          readOnly: true
        content:
          type: string
          maxLength: 1000
          description: Comment content
        postId:
          type: string
          format: uuid
          description: Associated post identifier
          readOnly: true
        authorId:
          type: string
          format: uuid
          description: Comment author identifier
          readOnly: true
        authorName:
          type: string
          description: Author display name
          readOnly: true
        parentId:
          type: string
          format: uuid
          nullable: true
          description: Parent comment ID for nested replies
          readOnly: true
        status:
          type: string
          enum: [active, moderated, deleted]
          description: Moderation status
          readOnly: true
        depth:
          type: integer
          minimum: 0
          description: Nesting level (0 for top-level comments)
          readOnly: true
        votes:
          type: object
          description: Vote counts
          readOnly: true
          properties:
            up:
              type: integer
              minimum: 0
              description: Upvote count
            down:
              type: integer
              minimum: 0
              description: Downvote count
            total:
              type: integer
              description: Net vote score
          required:
            - up
            - down
            - total
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          readOnly: true
        replies:
          type: array
          description: Nested reply comments (in threaded mode)
          readOnly: true
          items:
            $ref: '#/components/schemas/Comment'
      required:
        - id
        - content
        - postId
        - authorId
        - authorName
        - status
        - depth
        - votes
        - createdAt
        - updatedAt
...
    ModerationAction:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject, flag, delete]
          description: Moderation action to take
        reason:
          type: string
          maxLength: 500
          description: Reason for the action
        category:
          type: string
          enum: [spam, inappropriate, off-topic, harassment, other]
          description: Category of violation (for reject/flag actions)
        notifyUser:
          type: boolean
          default: true
          description: Whether to notify the comment author
      required:
        - action
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          description: Items per page
        totalPages:
          type: integer
          minimum: 0
          description: Total number of pages
        totalComments:
          type: integer
          minimum: 0
          description: Total number of comments
        hasNext:
          type: boolean
          description: Whether next page exists
        hasPrev:
          type: boolean
          description: Whether previous page exists
      required:
        - page
        - limit
        - totalPages
        - totalComments
        - hasNext
        - hasPrev
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        path:
          type: string
          description: API endpoint that generated the error
      required:
        - error
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid query parameters"
            code: "INVALID_PARAMETERS"
            timestamp: "2025-08-15T12:00:00Z"
            path: "/posts/123/comments"
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
            code: "UNAUTHORIZED"
            timestamp: "2025-08-15T12:00:00Z"
            path: "/posts/123/comments"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Insufficient permissions"
            code: "FORBIDDEN"
            timestamp: "2025-08-15T12:00:00Z"
            path: "/comments/123"
...
    NotFound:
      description: Comment not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Comment not found"
            code: "NOT_FOUND"
            timestamp: "2025-08-15T12:00:00Z"
            path: "/comments/123"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
            timestamp: "2025-08-15T12:00:00Z"
            path: "/posts/123/comments"
...